Move the WebMirrorStatus to internal package. Fix the list command of tunasynctl
